ccpm-tools/pyproject.toml
[project]
name = "ccpm-tools"
version = "0.1.5"
description = "CLI Python tools for CCPM."
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["typer-slim>=0.15", "click<8.2"]

[project.urls]
Homepage = "https://github.com/Deleranax/ccpm"
Issues = "https://github.com/Deleranax/ccpm/issues"

[project.scripts]
ccpm-tools = "main:app"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"